Prepare user messages for additional delivery surfaces
