+++ lib/features/user_message/user_message_host.dart
@@ -2,12 +2,11 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:lantern/core/common/app_dimens.dart';
-import 'package:lantern/core/common/app_semantic_colors.dart';
 import 'package:lantern/core/models/user_message.dart';
 import 'package:lantern/features/user_message/user_message_action_dispatcher.dart';
 import 'package:lantern/features/user_message/user_message_controller.dart';
 import 'package:lantern/features/user_message/user_message_route_observer.dart';
+import 'package:lantern/features/user_message/user_message_snackbar.dart';
 import 'package:loader_overlay/loader_overlay.dart';
 
 typedef UserMessageClock = DateTime Function();
@@ -25,9 +24,9 @@ class UserMessageHost extends ConsumerStatefulWidget {
     super.key,
   });
 
-  static const bodyKey = Key('user_message.snackbar.body');
-  static const actionKey = Key('user_message.snackbar.action');
-  static const closeKey = Key('user_message.snackbar.close');
+  static const bodyKey = UserMessageSnackbar.bodyKey;
+  static const actionKey = UserMessageSnackbar.actionKey;
+  static const closeKey = UserMessageSnackbar.closeKey;
 
   final Widget child;
   final UserMessageRouteObserver routeObserver;
@@ -172,86 +171,16 @@ class _UserMessageHostState extends ConsumerState<UserMessageHost>
 
   Widget _messageSurface(_UserMessagePresentation presentation) {
     final message = presentation.message;
-    final action = message.action;
-    return Positioned(
-      left: 16,
-      right: 16,
-      top: 0,
-      bottom: 16,
-      child: SafeArea(
-        top: false,
-        child: Align(
-          alignment: Alignment.bottomCenter,
-          child: SizedBox(
-            width: double.infinity,
-            child: Material(
-              key: ValueKey('user_message.snackbar.${message.displayId}'),
-              elevation: 6,
-              color: context.bgSnackbar,
-              clipBehavior: Clip.antiAlias,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(16),
-              ),
-              child: Padding(
-                padding: defaultPadding,
-                child: Row(
-                  children: [
-                    Expanded(
-                      child: ConstrainedBox(
-                        constraints: BoxConstraints(
-                          maxHeight: MediaQuery.sizeOf(context).height * 0.45,
-                        ),
-                        child: SingleChildScrollView(
-                          child: Semantics(
-                            liveRegion: true,
-                            label: message.body,
-                            excludeSemantics: true,
-                            child: Text(
-                              message.body,
-                              key: UserMessageHost.bodyKey,
-                              style: Theme.of(context).textTheme.bodyMedium
-                                  ?.copyWith(color: context.textInverse),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                    if (action != null && message.buttonLabel != null)
-                      KeyedSubtree(
-                        key: UserMessageHost.actionKey,
-                        child: TextButton(
-                          style: TextButton.styleFrom(
-                            foregroundColor: context.textInverseColor,
-                          ),
-                          onPressed: () {
-                            unawaited(widget.actionDispatcher.dispatch(action));
-                            _closePresentation(presentation);
-                          },
-                          child: Text(message.buttonLabel!),
-                        ),
-                      ),
-                    Semantics(
-                      label: MaterialLocalizations.of(
-                        context,
-                      ).closeButtonTooltip,
-                      button: true,
-                      child: ExcludeSemantics(
-                        child: IconButton(
-                          key: UserMessageHost.closeKey,
-                          icon: const Icon(Icons.close),
-                          color: context.textInverse,
-                          onPressed: () => _closePresentation(presentation),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
+    return switch (message.surface) {
+      UserMessageSurface.snackbar => UserMessageSnackbar(
+        message: message,
+        onAction: (action) {
+          unawaited(widget.actionDispatcher.dispatch(action));
+          _closePresentation(presentation);
+        },
+        onDismiss: () => _closePresentation(presentation),
       ),
-    );
+    };
   }
 
   void _onVisible(_UserMessagePresentation presentation) {
